+++ packages/pagination/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.12.39"></a>
+## [0.12.39](https://github.com/newsuk/times-components/compare/@times-components/pagination@0.12.38...@times-components/pagination@0.12.39) (2018-02-01)
+
+
+
+
+**Note:** Version bump only for package @times-components/pagination
+
 <a name="0.12.38"></a>
 ## [0.12.38](https://github.com/newsuk/times-components/compare/@times-components/pagination@0.12.37...@times-components/pagination@0.12.38) (2018-01-31)